refactor: Separate netwolf into subpackages
