+++ analysis_options.yaml
@@ -1,0 +1,9 @@
+include: package:very_good_analysis/analysis_options.yaml
+analyzer:
+  errors:
+    missing_enum_constant_in_switch: error
+linter:
+  rules:
+    always_put_required_named_parameters_first: false
+    flutter_style_todos: false
+    public_member_api_docs: false
