--- netwolf/lib/src/core/typedefs.dart
@@ -1,2 +1,0 @@
-typedef Id = int;
-typedef Json = Map<String, dynamic>;
