+++ netwolf/lib/src/extensions.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:netwolf/src/constants.dart';
-import 'package:netwolf/src/enums.dart';
+import 'package:netwolf_core/netwolf_core.dart';
 
 extension StringX on String {
   String toTitleCase() {
