+++ netwolf/lib/src/ui/widgets/settings_dialog.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:netwolf/src/constants.dart';
-import 'package:netwolf/src/core/netwolf_controller.dart';
-import 'package:netwolf/src/enums.dart';
 import 'package:netwolf/src/ui/widgets/base_dialog.dart';
+import 'package:netwolf_core/netwolf_core.dart';
 import 'package:notification_dispatcher/notification_dispatcher.dart';
 
 class SettingsDialog extends StatefulWidget {
