+++ netwolf/pubspec.yaml
@@ -1,23 +1,18 @@
 name: netwolf
 description: Inspired by Netfox on iOS. Netwolf allows for easy debugging of network requests made in your app via the use of interceptors.
 version: 0.3.0
-homepage: https://github.com/pongloongyeat/netwolf
+homepage: https://github.com/pongloongyeat/netwolf/netwolf
+publish_to: none
 
 environment:
   sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
-  collection: ^1.17.1
   flutter:
     sdk: flutter
+  netwolf_core:
+    path: ../netwolf_core
   notification_dispatcher: ^0.4.0
-  sqflite: ^2.3.0
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  mocktail: ^1.0.0
-  sqflite_common_ffi: ^2.3.0+2
   very_good_analysis: ^5.0.0+1
-
-flutter:
