+++ netwolf_core/lib/src/core/enums.dart
