+++ netwolf_core/lib/src/core/exceptions.dart
