+++ netwolf_core/lib/src/core/typedefs.dart
@@ -1,0 +1,1 @@
+typedef Id = int;
