+++ netwolf_core/lib/src/models/result.dart
