+++ netwolf_core/pubspec.yaml
@@ -1,0 +1,19 @@
+name: netwolf_core
+description: A starting point for Dart libraries or applications.
+version: 1.0.0
+homepage: https://github.com/pongloongyeat/netwolf/netwolf_core
+
+environment:
+  sdk: ">=3.0.0 <4.0.0"
+
+dependencies:
+  collection: ^1.17.1
+  meta: ^1.9.1
+  notification_dispatcher: ^0.4.0
+  sqflite: ^2.3.0
+
+dev_dependencies:
+  mocktail: ^1.0.0
+  sqflite_common_ffi: ^2.3.0+2
+  test: ^1.21.0
+  very_good_analysis: ^5.0.0+1
