+++ netwolf_tests/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 84a1e904f44f9b0e9c4510138010edcc653163f8
+  channel: stable
+
+project_type: package
