+++ netwolf_tests/test/netwolf_core/src/core/netwolf_controller_test.dart
@@ -2,9 +2,12 @@ import 'dart:math';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:netwolf/netwolf.dart';
-import 'package:netwolf/src/core/exceptions.dart';
-import 'package:netwolf/src/repositories/request_repository.dart';
+import 'package:netwolf_core/src/core/enums.dart';
+import 'package:netwolf_core/src/core/exceptions.dart';
+import 'package:netwolf_core/src/core/netwolf_controller.dart';
+import 'package:netwolf_core/src/models/netwolf_request.dart';
+import 'package:netwolf_core/src/models/result.dart';
+import 'package:netwolf_core/src/repositories/request_repository.dart';
 
 import '../../init.dart';
